+++ src/main/java/site/dogether/member/controller/MemberExceptionHandler.java
@@ -9,26 +9,32 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import site.dogether.common.controller.response.ApiResponse;
 import site.dogether.member.controller.response.MemberExceptionCode;
 import site.dogether.member.domain.exception.InvalidMemberException;
-import site.dogether.member.exception.MemberNotFoundException;
+import site.dogether.member.service.exception.MemberNotFoundException;
 
 @Slf4j
 @Order(Ordered.HIGHEST_PRECEDENCE)
 @RestControllerAdvice
 public class MemberExceptionHandler {
 
-    @ExceptionHandler(InvalidMemberException.class)
+    @ExceptionHandler
     public ResponseEntity<ApiResponse<Void>> handleInvalidMemberException(final InvalidMemberException e) {
         log.warn(e.getMessage());
 
         return ResponseEntity.badRequest()
-                .body(ApiResponse.fail(MemberExceptionCode.INVALID_MEMBER_EXCEPTION, e.getMessage()));
+                .body(ApiResponse.fail(MemberExceptionCode.INVALID_MEMBER,
+                    e.getMessage()));
     }
 
-    @ExceptionHandler(MemberNotFoundException.class)
+    @ExceptionHandler
     public ResponseEntity<ApiResponse<Void>> handleMemberNotFoundException(final MemberNotFoundException e) {
         log.warn(e.getMessage());
 
+        /**
+         * TODO : Error Code에 메시지를 넣게 된다면 fail 메서드 오버로딩
+         */
         return ResponseEntity.badRequest()
-                .body(ApiResponse.fail(MemberExceptionCode.MEMBER_NOT_FOUND_EXCEPTION, e.getMessage()));
+                .body(ApiResponse.fail(
+                    MemberExceptionCode.MEMBER_NOT_FOUND,
+                    MemberExceptionCode.MEMBER_NOT_FOUND.getMessage()));
     }
 }
