+++ src/main/java/site/dogether/member/controller/response/MemberExceptionCode.java
@@ -9,9 +9,9 @@ import site.dogether.common.exception.ExceptionCode;
 @RequiredArgsConstructor(access = AccessLevel.PACKAGE)
 public enum MemberExceptionCode implements ExceptionCode {
 
-    INVALID_MEMBER_EXCEPTION("MF-0001"),
-    MEMBER_NOT_FOUND_EXCEPTION("MF-0002");
+    INVALID_MEMBER("MF-0001", "유효하지 않은 회원 정보입니다."),
+    MEMBER_NOT_FOUND("MF-0002", "존재하지 않는 회원 정보입니다.");
 
     private final String value;
-
+    private final String message;
 }
